+++ cli/integration-test/sample-test/src/sample_test/watched_files_change/a.clj
@@ -1,0 +1,5 @@
+(ns sample-test.watched-files-change.a)
+
+(defn foo [] 1)
+
+(foo)